Move template.js to the Site folder
